requirements.txt
colorlog
thespian
PyYAML
pretty-errors
faiss-cpu
numpy
regex
tqdm
langdetect
googletrans==3.1.0a0
translate
clip-by-openai
torch
pillow
scipy
matplotlib
pandas
tqdm
salesforce-lavis